[2.0-fixes] rename sample package
