+++ sample/AirshipSample/android/app/src/main/java/com/urbanairship/sample/MainActivity.java
@@ -1,4 +1,4 @@
-package com.airshipsample;
+package com.urbanairship.sample;
 
 import com.facebook.react.ReactActivity;
